+++ src/main/java/com/lifeonwalden/codeGenerator/constant/DefinedMappingID.java
@@ -1,0 +1,5 @@
+package com.lifeonwalden.codeGenerator.constant;
+
+public interface DefinedMappingID {
+    String BASE_RESULT_MAP = "baseResultMap";
+}
