+++ src/main/java/com/lifeonwalden/codeGenerator/constant/SpecialInnerSuffix.java
@@ -1,0 +1,18 @@
+package com.lifeonwalden.codeGenerator.constant;
+
+public interface SpecialInnerSuffix {
+
+    String PICKED = "Picked";
+
+    String START = "Start";
+
+    String END = "End";
+
+    String LIKE = "Like";
+
+    String NOT_LIKE = "NotLike";
+
+    String IN = "In";
+
+    String NOT_IN = "NotIn";
+}
