+++ src/main/java/com/lifeonwalden/codeGenerator/mybatis/impl/XMLMapperGenerator.java
@@ -3,6 +3,12 @@ package com.lifeonwalden.codeGenerator.mybatis.impl;
 import com.lifeonwalden.codeGenerator.bean.Table;
 import com.lifeonwalden.codeGenerator.bean.config.Config;
 import com.lifeonwalden.codeGenerator.mybatis.DomGenerator;
+import com.lifeonwalden.codeGenerator.mybatis.impl.column.BaseColumnListElementGenerator;
+import com.lifeonwalden.codeGenerator.mybatis.impl.column.BaseColumnListWithPrefixElementGenerator;
+import com.lifeonwalden.codeGenerator.mybatis.impl.column.ColumnListElementXcludeDBFieldGenerator;
+import com.lifeonwalden.codeGenerator.mybatis.impl.column.ColumnListWithPrefixElementXcludeDBFieldGenerator;
+import com.lifeonwalden.codeGenerator.mybatis.impl.resultmap.BaseResultMapElementGenerator;
+import com.lifeonwalden.codeGenerator.mybatis.impl.resultmap.ResultMapElementXcludeDBFieldGenerator;
 import com.lifeonwalden.codeGenerator.util.StringUtil;
 import org.mybatis.generator.dom.constant.XmlConstants;
 import org.mybatis.generator.dom.xml.Document;
