+++ src/main/java/com/lifeonwalden/codeGenerator/util/StringUtil.java
@@ -1,14 +1,14 @@
 package com.lifeonwalden.codeGenerator.util;
 
-public class StringUtil {
-    public static String firstAlphToUpper(String string) {
+public interface StringUtil {
+    static String firstAlphToUpper(String string) {
         char[] buffer = string.toCharArray();
         buffer[0] = Character.toUpperCase(buffer[0]);
 
         return new String(buffer);
     }
 
-    public static String removeUnderline(String string) {
+    static String removeUnderline(String string) {
         String[] words = string.split("_");
         int length = words.length;
 
@@ -23,14 +23,16 @@ public class StringUtil {
         return builder.toString();
     }
 
-    public static void main(String[] args) {
-        System.out.println(firstAlphToUpper("aere123"));
-        System.out.println(firstAlphToUpper("Aere123"));
-        System.out.println(firstAlphToUpper("1ere123"));
-
-        System.out.println(removeUnderline("1ere123"));
-        System.out.println(removeUnderline("1ere_123"));
-        System.out.println(removeUnderline("ABC_DEFGHIJKLMN"));
-        System.out.println(removeUnderline("ABC_DEFGHI_JKLMN"));
+    static boolean isNotBlank(String string) {
+        int strLen;
+        if (string == null || (strLen = string.length()) == 0) {
+            return false;
+        }
+        for (int i = 0; i < strLen; i++) {
+            if (Character.isWhitespace(string.charAt(i)) == false) {
+                return true;
+            }
+        }
+        return false;
     }
 }
